requirements.txt
redis
# "uvicorn[standard]"
fastapi
requests
httpie
serpapi
transformers
torch
pillow